chore: Update getCookies function to separate file
