+++ apps/earn-protocol/constants/get-cookies.ts
@@ -1,0 +1,5 @@
+export const getCookies = (cookiename: string) => {
+  const cookiestring = RegExp(`${cookiename}=[^;]+`, 'u').exec(document.cookie)
+
+  return decodeURIComponent(cookiestring ? cookiestring.toString().replace(/^[^=]+./u, '') : '')
+}
