+++ apps/earn-protocol/constants/safe-parse-json.ts
@@ -1,0 +1,10 @@
+export const safeParseJson = (json: string) => {
+  try {
+    return JSON.parse(json)
+  } catch (error) {
+    // eslint-disable-next-line no-console
+    console.error('Error parsing JSON', error)
+
+    return {}
+  }
+}
